Fixes SocketCluster/socketcluster#142
Previously, the init and broker controllers would be require'd immediately. However, if a user was wanting to address alternative dialects such as ES6 or CoffeeScript, they would fail and would have to drop the registrars outside of the function body.

This, unfortunately, is asymmetric with how SocketCluster's core cluster actually works. There, the controllers are instantiated and run in order, allowing the coder to focus just on the `.run` method.

This here fixes that. The code should now be symmetric and predictable.